Added consideration of vendors when using header based version selection
class MyAPI < Grape::API
vendor 'v1', :using => :header, :vendor => 'twitter', :format => :json

get '/' do
'hello world'
end
end

and let it still return the '/' even when the accept-header
"application/vnd.some_other_vendor-v1+json" was given. This commit will
let it return a 404, and sets the X-Cascade pass as well.

I've added this because although the version is right, part of the
Accept header (the vendor part) is not correct.