Grape::API versioning based on request parameter.
Added an API versioner which searches for a named parameter (default is
'apiver') to find the specified api version. The name of the parameter
can be changed via the 'parameter' option when specifying the version.

Include all the same specs as already existed for the other versioners.
Update README markdown to explain this new versioning strategy.