adds 405 response for methods not supported on a resource instead of 404
For every resource an 'OPTIONS' route that returns an HTTP 204 response will
be added. Also a route that will return an HTTP 405 response for
any HTTP method that the resource cannot handle will be added.

The behavior up to now has been to return an HTTP 404 response if the
resource doesn't respond to the request's HTTP method.