Extend :cascade to work with all versioning strategies
The :cascade option allows preventing X-Cascade to be returned on
response headers. For Rails users, this is useful because your
API exceptions don't get rendered/handled by Rails, but actually sent to
users. This is particularly useful if you want your user to receive an
error exception in the same format your API supports, and not an
unexpected HTML page.

A more in-depth explanation of :cascade was moved into the code and the
overview was inserted on the main README on the exceptions section.

This commit also changes the order in which versioning strategies are
presented. The order is now consistent with the order in which the
strategies were first presented. It also presents the default one,
before introducing non-defaults.