Remove a mysterious zero-width replacement
As far as I can tell, it replaced zero-width breakers with zero-width
replacement (that is, it seemed to have no effect).

No tests failed when I removed it. If it does do something, maybe a test or
a comment would be handy to explain the intention.

It was introduced in commit f7728b88c8cb33e5a783f7a779c761196d697343 -- the
original implementation.