Add test to reproduce bug with Warden test helper.
When using the `login_as` Warden test helper, it depends whether or not
the API class under test is evaluated *before* or *after* the tests.

Given that this bug first appeared in v0.10.0, my current theory is the
use of `deep_dup` is responsible. Warden persists registered hooks in
instance variables (ex. http://git.io/NhKG). When an API class is
evaluated *before* the tests, the hooks are registered *after*
duplication takes place, so the original object is left with an empty
array of hooks.