Add failing test for apparent route_param bug.
When a Grape::API is set to support just a single format (e.g. `format :json`), it does not need URL extensions to specify the format of a request. Under such conditions, it should be possible to pass a string value that contains a period (.) to a `route_param`, which would otherwise be recognized as a format extension.