A possible fix/conversation starter for #930.
@dblock's suggestion (http://git.io/vvmJj) was instead of `deep_dup`ing
_all_ settings by default, to only `deep_dup` settings on a whitelist.
However, after implementing this, at least according to the tests, there
doesn't appear to be **any** settings need to be `deep_dup`ed. Instead,
a shallow clone (a `dup`) looks to be sufficient.

I've confirmed that with this change, the test
`warden_test_helpers_spec.rb` passes in the suite and in isolation (i.e.
the bug described is fixed).