Support custom types.
Any class can now be used as a type for a parameter, so long
as it defines a class-level `parse` method. This method
should raise on invalid values and otherwise return the coerced
value.

Also, I added the Grape::ParameterTypes module, which serves
right now only as a basis to determine if a given type is a custom
type or not, but in the future could be used to actually validate
the types that are given via the params DSL.