Fix 500 error for xml format when method is not allowed
Problem:

All requests to Not Allowed methods are returning 500 error instead of
405 when the format is xml.

Cause:

Route definitions for not allowed methods are returning an empty string
as a response body. But the empty string fails to convert to xml and
raises InvalidFormatter error.

Solution:

Define an exception for "method is not allowed" case.
Return an error message "405 Not Allowed" instead of empty string.