Fix infinite loop when mounting endpoint with same superclass
Subclass APIs were being assigned the same top level inheritable
settings from the superclass. When an API attempted to mount another
API with the same superclass, the same setting object would be added to
the setting stack. When an attempt was made to retrieve a value from the
settings, it would sit in loop calling itself.