Fix exception being thrown when dependent parameter is a Hash
When using `given` and the paramter type is a Hash,
a Grape::Exceptions::UnknownParameter is thrown. This is a wrong behavior
due to the fact that the method `declared_param?(param)` do not deal
properly with params that are hashes (even after flattening the
declared params).

This commit fixes this by checking if the param is a Hash and thus
returning its first key. This approach was already used in other part of
the code, so it was extracted and reused on `declared_params?(param)`.