Allow a user to insert a middleware at a specific location in the stack.
This commit adds a documented but not implemented feature to allow a user to insert a middleware at a specific point in the stack. It does this by extending the Grape::DSL::Middleware class with a new method `.insert` which delegates down to the already existing methods in existance on the Grape::Middleware::Stack class.