Use defined? to correct not initialized warnings
The class instances variables `@versions` and `@namespace_description` may not be initialized when used and Ruby will report their access as a warning. Using the defined? keyword will correct that. There are plenty of alternative changes could be made to avoid this, but I just picked the most minimal change.