replace Virtus with dry-types
As we know Virtus isn't maintained anymore, so it makes sense to use something
different for coercing. There is a dry-types lib which was created as
a replacement for Virtus (at least a part of it). Although, it is only
about coercion.

This change gets rid of Virtus and adds dry-types to the stack.

The structure inside didn't change match. Dry-types was integrated
without huge refactoring.

If people want to use Virtus for custom types, it is possible after
implementing a parse method which initializes a model.

      class User
        include Virtus.model

        attribute :id, Integer
        attribute :name, String

        def self.parse(*args)
          new(*args)
        end
      end