introduce a finalize! method
There are a few variables (settings) which are only needed
while defining API, once everything is set up those variables
could be cleaned up to save memory.

So, the `finanlize!` method was added to the API and Instance classes,
I found only few variables which might be cleaned up, over time
there might be more.

The `finalize!` method could be a part of the `compile!` method, however,
`compile!` must be called on a child level (for example, `Twitter::API.compile!`)
while `finalize!` must be called on `Grape::API` to find all inheritors
and clean them up.