Expand test coverage of undocumented functionality
In an interview recently, I was given the Grape repository sans the
content of this method as a technical screening question. The task was
to discover what had been removed from the project and re-implement the
missing code using only the test suite.

It was actually fairly difficult to re-implement this method because
there are a couple of untested responsibilities. I think this is a
fairly integral component of the framework, so it makes sense to fully
test its functionality.