a setting for disabling documentation to internal APIs
Our application has 3 APIs, 2 of them are internal and don't require
documentations. However, there has not been any option to prevent
Grape from documenting parameters for internal APIs.

This change adds a `do_not_document!` setting which instructs Grape
to not document parameters, thus needless objects allocation is avoided.

    class Api < Grape::API
      do_not_document!
    end

The logic for documenting parameters was moved to a separate class,
the `Grape::Validations::ParamsScope` class has to many responsibilities,
it would be better to split it.