Move convenience class into new file
This commit moves the `Grape::Types::InvalidValue` class definition into
a separate file so that it can be lazy loaded in projects that use
zeitwerk file naming conventions.

In environments that do not eager load, referencing the
`Grape::Validations::Types::InvalidValue` class before
`Grape::Types::InvalidValue` causes a load error due to the fact that
both classes were defined in the same file.